PenciWriteAdjustsDurabilityCorrectilyForWhiteSpaceWithMoreLettersThanDurabilityLeft Green (No Red commit)
Bad on me. This test did not fail first try, so was it a valid test to run? Or did I just write too much code and not the bare minimum of the previous commit.

I'm on the side that I wrote too much code (return "T   " would have worked, if tipDurability would be < 0 ) so I refactored and did a few things too early for pure TDD.

I suppose we all get carried away sometimes.